Eigene Durchsicht: 25 Fehler in Threads, DSP-Numerik und Parsern
Drei getrennte Lesedurchgaenge (Plugin-Schicht, Signalpfad, Parser und
Lader), jeder Fund am Code verifiziert, jeder Fix mit Test oder Nachweis.

Threads
- releaseIdleEngine las fading_ VOR swapTo_. Der Audio-Thread schreibt
  fading_ und loescht swapTo_ zuletzt; in der Luecke zwischen den beiden
  Loads sah ein gerade begonnener Crossfade aus wie Stillstand, und die
  Funktion gab die Engine frei, die der Audio-Thread rendert. Jetzt
  swapTo_ zuerst (an drei Stellen).
- playSetFile tauschte die Ereignisliste, waehrend der Audio-Thread darin
  stehen konnte: setPlayBusy_-Handschlag wie bei der Aufnahme.
- Doppelpuffer-Uebergabe (Texturen, Impulsantworten): der Warte-Spin war
  200 000 Umdrehungen, ein Fuenftel einer Millisekunde -- kuerzer als
  jeder Block. Danach wurde der Puffer unter dem Leser neu belegt. Jetzt
  Schlafwarten bis 200 ms.
- OSC-Thread wird vor der Warteschlange gestoppt, in die er schreibt.
- heldVel_, morphSelectSeconds_ atomar. Zwei Tabellen (eingebaute
  Wavetables, Z-Plane-Gitter) wurden erstmals im Audio-Thread gebaut,
  unter der Sperre einer funktionslokalen Static: prepare fasst sie an.

Signalpfad
- Patina und Voice-Comb rechneten die Leseposition als float(w_). Ab
  2^24 Samples -- sechs Minuten bei 48 kHz -- ist das keine Ganzzahl
  mehr, und der Lesekopf jitterte gegen den Schreibzeiger, auf dem
  Master. Jetzt double, Zeiger maskiert.
- zModal_.prepare() wurde nie gerufen: der Modalfilter rechnete immer
  mit 48 kHz, bei 96 kHz eine Oktave zu hoch. Nachweis: Centroid 247 Hz
  bei 96 kHz gegen 249 Hz bei 44,1 kHz.
- fb-Tape-Ring war zwei Bloecke lang, das Wow liest weiter zurueck.
  Diffuser-Ring fest 8192 Samples, kuerzer als seine 47,9-ms-Stufe ab
  176,4 kHz. silentSamples_ war long (32 Bit unter MSVC, Wrap nach 12 h).
- Sympathy: 16 Stimmen mal 0,12 mal Filterresonanz kann eine Schleifen-
  verstaerkung ueber eins ergeben; nichts begrenzte sie. Der Koppelpuffer
  ist jetzt auf plus/minus 2 geklemmt -- saettigen ja, nach Inf laufen
  nein. Stretch-wrap() lief in while-Schleifen O(q/loopLen): weit weg
  jetzt fmod, nahe unveraendert.
- Early Room: Ohr- und Wandpaar-Distanzen gleiten jetzt wie die Quelle,
  statt bei Size-Automation je Block zu springen.

Parser und Lader
- "nan" passiert jeden min/max-Vergleich: Parameter aus Text, Mod-Depth,
  Huellkurvenpunkte, Set-Zeiten (hielten die Wiedergabe fuer immer an),
  Score-Werte. Ueberall isnan-Pruefung; inf klemmen bleibt.
- Stapelmodus: loadDefaultPresetPacks fiel bei n == 0 im AMBIENT_PACKS-
  Zweig auf die Standardordner durch -- und beim zweiten Aufruf im selben
  Prozess ist n immer 0. Ein Batch las ab Eintrag 2 die installierten
  Packs aus ProgramData nach. Die Messung war nicht betroffen (gleiche
  Packnamen, am Namen abgewiesen); Nachweis mit Ein-Pack-Ordner. Pfad
  gilt erst als geladen, wenn er es ist.
- WAV: data-Chunk-Laenge 0xFFFFFFFF (abgerissene Aufnahme) forderte
  16 GB an. Jetzt auf die Restdateigroesse gekappt; Testdatei laedt.
- snprintf-Soll-Laenge als Kopierlaenge (Set- und Routen-Text).
- Scala: leere Beschreibungszeile wurde uebersprungen, "3/2 pure fifth
  (approx. 702c)" als Cents gelesen. Selbsttest fuer beides.
- Renderer: --block 0 war eine Endlosschleife, --seconds -1 ein Wurf.
- CC 74 war nie lernbar (eigener else-if vor dem Controller-Zweig).
  currentProgram_ blieb nach setStateInformation stehen.

Klang unveraendert: derselbe Render-Hash der Referenz. Selbsttest,
Hosttest, Standalone-Uebergang gruen.

+++ Core/include/ambient/Engine.h
@@ -566,7 +566,7 @@ private:
     // per two seconds so that a press is a slide and never a jump (SOMA Terra's interval keys).
     double            transposeTarget_ = 0.0, transposeCur_ = 0.0;
     // Sleep: after two seconds of silence (no voice, output below -90 dBFS) the effects sleep
-    long              silentSamples_ = 0;
+    long long         silentSamples_ = 0;   // 64-bit: a long is 32 bits here and wrapped after twelve hours
     bool              asleep_ = false;
     // The map cursor the blend target was computed for, so a cursor that stands still is not
     // searched for and blended again on every block.
